EntityRedDoor matching everywhere else
