+++ config/splat.us.stcen.yaml
@@ -38,7 +38,8 @@ segments:
       - [0x1300, data]
       - [0xD42C, rodata]
       - [0xD4B0, .rodata, D600]
-      - [0xD4D8, rodata]
+      - [0xD4D8, .rodata, 11280] # EntityRedDoor
+      - [0xD4F0, rodata]
       - [0xD578, .rodata, entity_relic_orb]
       - [0xD5B0, .rodata, entity_message_box]
       - [0xD5C4, .rodata, 18084] # EntityStageNamePopup
