+++ config/splat.us.stdre.yaml
@@ -50,7 +50,8 @@ segments:
       - [0x1181C, rodata]
       - [0x118D0, rodata]
       - [0x11910, .rodata, 1697C] # EntityFadeToWhite
-      - [0x1193C, rodata]
+      - [0x1193C, .rodata, 173C4] # EntityRedDoor
+      - [0x11954, rodata]
       - [0x119C4, .rodata, 1C7DC]
       - [0x119DC, .rodata, entity_relic_orb]
       - [0x11A14, .rodata, entity_message_box]
