+++ config/splat.us.stno3.yaml
@@ -64,7 +64,8 @@ segments:
       - [0x374F4, .rodata, 3FF00] # EntityDeathStolenItems
       - [0x37514, .rodata, 3FF00] # EntityDeath
       - [0x37564, .rodata, 3FF00] # EntityWargExplosionPuffOpaque
-      - [0x37578, rodata]
+      - [0x37578, .rodata, 41C80] # EntityRedDoor
+      - [0x37590, rodata]
       - [0x375E0, .rodata, 46684] # EntityPrizeDrop
       - [0x37600, .rodata, 46684] # EntityEquipItemDrop
       - [0x37618, .rodata, entity_relic_orb]
