+++ config/symbols.beta.txt
@@ -103,6 +103,6 @@ SetGeomOffset = 0x8001803C;
 
 D_8003BE1E_mad = 0x8003BE1E;
 D_8003BF28_mad = 0x8003BF28;
-D_80072E9C_mad = 0x80072E9C;
-D_80072EB4_mad = 0x80072EB4;
+PLAYER_zPriority = 0x80072E9C;
+PLAYER_step = 0x80072EB4;
 D_80096ED0_mad = 0x80096ED0;
