+++ config/symbols.stmad.txt
@@ -13,6 +13,8 @@ g_eBreakableanimSets = 0x80180674;
 g_eBreakableDrawModes = 0x80180684;
 g_eDamageDisplayClut = 0x801809C4;
 g_Rooms = 0x80181130;
+g_eRedDoorUV = 0x801809D4;
+g_eRedDoorTiles = 0x80180594;
 g_pStrObjLayout0 = 0x801811D4;
 g_pStrObjLayout1 = 0x801811E8;
 g_pStrObjLayout2 = 0x80181206;
@@ -69,7 +71,6 @@ g_InitializeData0 = 0x801804FC;
 g_UnkPrimHelperRot = 0x8018D8C0;
 UnkPrimHelper = 0x80198FA0;
 EntityUnkId15 = 0x80195B44;
-EntityIsNearPlayer = 0x801910A8;
 UnkPolyFunc2 = 0x801996EC;
 PrimDecreaseBrightness = 0x8019976C;
 g_LayoutObjHorizontal = 0x801997D8;
