+++ config/symbols.us.stcen.txt
@@ -7,8 +7,10 @@ g_InitializeEntityData0 = 0x8018041C;
 g_EInitGeneric = 0x80180434;
 g_eInitGeneric2 = 0x80180458;
 g_eDamageDisplayInit = 0x80180464;
+g_eRedDoorTiles = 0x801804A4;
 UNK_Invincibility0 = 0x801807B8;
 g_eDamageDisplayClut = 0x80180B94;
+g_eRedDoorUV = 0x80180BA4;
 c_GoldPrizes = 0x80180F58;
 c_HeartPrizes = 0x80180FE8;
 g_Rooms = 0x801812D4;
@@ -31,7 +33,6 @@ EntityDamageDisplay = 0x80192398;
 InitRoomEntities = 0x80193298;
 UpdateRoomPosition = 0x80193410;
 CreateEntityFromEntity = 0x80193538;
-EntityIsNearPlayer = 0x801935B4;
 EntityRedDoor = 0x8019362C;
 DestroyEntity = 0x80194264;
 PreventEntityFromRespawning = 0x8019434C;
