+++ config/symbols.us.stdre.txt
@@ -8,6 +8,7 @@ g_InitializeEntityData0 = 0x80180470;
 g_EInitGeneric = 0x80180488;
 g_eInitGeneric2 = 0x801804AC;
 g_eDamageDisplayInit = 0x801804B8;
+g_eRedDoorTiles = 0x80180550;
 g_eBreakableAnimations = 0x80180610;
 g_eBreakableHitboxes = 0x80180630;
 g_eBreakableExplosionTypes = 0x80180638;
@@ -16,6 +17,7 @@ g_eBreakableDrawModes = 0x80180650;
 g_CloneShootOrder = 0x801807D4;
 UNK_Invincibility0 = 0x8018097C;
 g_eDamageDisplayClut = 0x80180D58;
+g_eRedDoorUV = 0x80180D68;
 c_GoldPrizes = 0x8018111C;
 c_HeartPrizes = 0x801811AC;
 g_Rooms = 0x80181498;
@@ -44,12 +46,10 @@ InitRoomEntities = 0x801993DC;
 UpdateRoomPosition = 0x80199554;
 CreateEntityFromCurrentEntity = 0x80199608;
 CreateEntityFromEntity = 0x8019967C;
-EntityIsNearPlayer = 0x801996F8;
 EntityRedDoor = 0x80199770;
 DestroyEntity = 0x8019A3A8;
 PreventEntityFromRespawning = 0x8019A490;
 AnimateEntity = 0x8019A4D8;
-GetSideToPlayer = 0x8019A718;
 MoveEntity = 0x8019A75C;
 FallEntity = 0x8019A78C;
 AllocEntity = 0x8019AC18;
