+++ config/symbols.us.stno3.txt
@@ -8,6 +8,7 @@ g_InitializeEntityData0 = 0x80180AC4;
 g_EInitGeneric = 0x80180ADC;
 g_eInitGeneric2 = 0x80180B00;
 g_eDamageDisplayInit = 0x80180B0C;
+g_eRedDoorTiles = 0x80180DA0;
 g_eBreakableAnimations = 0x80180E50;
 g_eBreakableHitboxes = 0x80180E70;
 g_eBreakableExplosionTypes = 0x80180E78;
@@ -16,6 +17,7 @@ g_eBreakableDrawModes = 0x80180E90;
 g_TrapDoorFlag = 0x8018123C;
 UNK_Invincibility0 = 0x80181CC0;
 g_eDamageDisplayClut = 0x8018209C;
+g_eRedDoorUV = 0x801820AC;
 c_GoldPrizes = 0x80182460;
 c_HeartPrizes = 0x801824F0;
 g_Rooms = 0x80183CC4;
@@ -76,7 +78,6 @@ InitRoomEntities = 0x801C3C98;
 UpdateRoomPosition = 0x801C3E10;
 CreateEntityFromCurrentEntity = 0x801C3EC4;
 CreateEntityFromEntity = 0x801C3F38;
-EntityIsNearPlayer = 0x801C3FB4;
 EntityRedDoor = 0x801C402C;
 DestroyEntity = 0x801C4C64;
 PreventEntityFromRespawning = 0x801C4D4C;
