+++ config/symbols.us.stnp3.txt
@@ -8,6 +8,7 @@ g_InitializeEntityData0 = 0x80180A54;
 g_EInitGeneric = 0x80180A6C;
 g_eInitGeneric2 = 0x80180A90;
 g_eDamageDisplayInit = 0x80180A9C;
+g_eRedDoorTiles = 0x80180D78;
 g_eBreakableAnimations = 0x80180E28;
 g_eBreakableHitboxes = 0x80180E48;
 g_eBreakableExplosionTypes = 0x80180E50;
@@ -16,6 +17,7 @@ g_eBreakableDrawModes = 0x80180E68;
 g_TrapDoorFlag = 0x80181104;
 UNK_Invincibility0 = 0x8018164C;
 g_eDamageDisplayClut = 0x80181A28;
+g_eRedDoorUV = 0x80181A38;
 c_GoldPrizes = 0x80181DEC;
 c_HeartPrizes = 0x80181E7C;
 g_Rooms = 0x80183A7C;
@@ -53,7 +55,6 @@ InitRoomEntities = 0x801BB508;
 UpdateRoomPosition = 0x801BB680;
 CreateEntityFromCurrentEntity = 0x801BB734;
 CreateEntityFromEntity = 0x801BB7A8;
-EntityIsNearPlayer = 0x801BB824;
 EntityRedDoor = 0x801BB89C;
 DestroyEntity = 0x801BC4D4;
 PreventEntityFromRespawning = 0x801BC5BC;
