+++ config/symbols.us.stnz0.txt
@@ -10,12 +10,14 @@ g_InitializeEntityData0 = 0x80180BE0;
 g_EInitGeneric = 0x80180BF8;
 g_eInitGeneric2 = 0x80180C1C;
 g_eDamageDisplayInit = 0x80180C28;
+g_eRedDoorTiles = 0x80180D8C;
 g_CallElevator = 0x80180EEC;
 g_BossFlag = 0x80181010;
 g_ElevatorTarget = 0x801813A4;
 g_ElevatorTargetPos = 0x801813A8;
 UNK_Invincibility0 = 0x80181574;
 g_eDamageDisplayClut = 0x80181950;
+g_eRedDoorUV = 0x80181960;
 c_GoldPrizes = 0x80181D14;
 c_HeartPrizes = 0x80181DA4;
 g_Rooms = 0x8018272C;
@@ -52,7 +54,6 @@ InitRoomEntities = 0x801BB920;
 UpdateRoomPosition = 0x801BBA98;
 CreateEntityFromCurrentEntity = 0x801BBB4C;
 CreateEntityFromEntity = 0x801BBBC0;
-EntityIsNearPlayer = 0x801BBC3C;
 EntityRedDoor = 0x801BBCB4;
 DestroyEntity = 0x801BC8EC;
 PreventEntityFromRespawning = 0x801BC9D4;
