+++ config/symbols.us.stst0.txt
@@ -6,10 +6,12 @@ g_InitializeData0 = 0x80180580;
 g_InitializeEntityData0 = 0x8018058C;
 g_eInitGeneric2 = 0x801805BC;
 g_eDamageDisplayInit = 0x801805C8;
+g_eRedDoorTiles = 0x8018064C;
 g_isSecretStairsButtonPressed = 0x8018089C;
 g_isDraculaFirstFormDefeated = 0x8018090C;
 UNK_Invincibility0 = 0x801815EC;
 g_eDamageDisplayClut = 0x80181968;
+g_eRedDoorUV = 0x80181978;
 g_Rooms = 0x80182060;
 D_80183D1C = 0x80183D1C;
 D_80197F80 = 0x80197F80;
@@ -42,7 +44,6 @@ InitRoomEntities = 0x801B393C;
 UpdateRoomPosition = 0x801B3AB4;
 CreateEntityFromCurrentEntity = 0x801B3B68;
 CreateEntityFromEntity = 0x801B3BDC;
-EntityIsNearPlayer = 0x801B3C58;
 EntityRedDoor = 0x801B3CD0;
 DestroyEntity = 0x801B4908;
 DestroyEntitiesFromIndex = 0x801B4974;
