+++ config/symbols.us.stwrp.txt
@@ -50,13 +50,10 @@ Random = 0x801881E8;
 TestCollisions = 0x8018861C;
 EntityDamageDisplay = 0x80189734;
 CreateEntityFromEntity = 0x8018A8D4;
-EntityIsNearPlayer = 0x8018A950;
 EntityRedDoor = 0x8018A9C8;
 DestroyEntity = 0x8018B600;
 PreventEntityFromRespawning = 0x8018B6E8;
-GetSideToPlayer = 0x8018B970;
 AllocEntity = 0x8018BE70;
-InitializeEntity = 0x8018C310;
 CollectGold = 0x8018CD10;
 EntityEquipItemDrop = 0x8018DA34;
 EntityStageNamePopup = 0x8019198C;
