+++ include/sfx.h
@@ -193,7 +193,15 @@ typedef enum { MONO_SOUND, STEREO_SOUND } soundMode;
 #define NA_SE_EV_GLASS_BREAK 0x632
 #define NA_SE_BREAK_CANDLE 0x634
 #define NA_SE_PL_WARP 0x636
+
+#if defined(VERSION_BETA)
+#define SFX_OPEN_DOOR 0x640
+#define SFX_DOOR_UNKNOWN 0x64D
+#else
 #define SFX_OPEN_DOOR 0x642
+#define SFX_DOOR_UNKNOWN 0x64F
+#endif
+
 #define NA_SE_EN_COG_CLICK 0x642
 #define NA_SE_EN_ROCK_BREAK 0x644
 #define SOUND_BAT_SCREECH 0x64E
