+++ src/st/get_side_to_player.h
@@ -4,7 +4,7 @@
  * 1 = Player is on the left side
  * 2 = Player is above
  */
-s16 GetSideToPlayer(void) {
+s32 GetSideToPlayer(void) {
     s16 var_a0 = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
 
     if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
