+++ src/st/st0/st0.h
@@ -28,7 +28,7 @@ typedef enum {
 } EntityIDs;
 
 void CreateEntityFromEntity(u16 entityId, Entity* source, Entity* entity);
-s16 GetSideToPlayer();
+s32 GetSideToPlayer();
 void MoveEntity();
 void SetStep(u8);
 void EntityExplosion(Entity*);
